refactor(#1526): extract persistExtraction to PersistCoordinator (seam 16)
Move the 2.5k-line persistExtraction method from orchestrator.ts into a
new PersistCoordinator (orchestration/extraction-persist.ts). This is the
largest single orchestrator-resident method, handling:
- inline citation injection, redaction gating (#1669)
- extraction-judge gating (#562), importance gating
- scope routing (global → shared namespace)
- 3-path dedup (hash, semantic, exact)
- bi-temporal backfill on dedup hits (#1671/#1707)
- tombstone postWriteGuard (#1645)
- promotion paths (profile/shared), content-hash indexing
- contradiction auto-resolve deferral, link suggestion, graph edges
- behavior-signal tracking, temporal/tag index updates

Behavior-preserving move via deps-delegate + lazy-getter pattern
(seams 13-15 playbook). Orchestrator.ts: 18035 → 15524 LOC (-2511).

Refs #1526
